Kwikstick-K40: Rename files for better conformance to naming conventions
